requirements.txt
absl-py==1.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
attrs==24.2.0
auto-py-to-exe==2.42.0
based58==0.1.1
bitarray==3.0.0
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.3
cassandra-driver==3.28.0
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
colorama==0.4.6
cql==1.4.0
cryptography==43.0.1
customtkinter==5.2.1
cytoolz==1.0.0
darkdetect==0.8.0
decorator==5.1.1
distlib==0.3.8
dm-tree==0.1.8
dnspython==2.6.1
ecdsa==0.18.0
Eel==0.16.0
et-xmlfile==1.1.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
etils==1.7.0
filelock==3.13.3
fonttools==4.46.0
frozenlist==1.5.0
fsspec==2024.3.0
future==0.18.3
geocoder==1.38.1
geomet==0.2.1.post1
gevent==23.9.1
gevent-websocket==0.10.1
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
googletrans==3.0.0
greenlet==3.0.2
h11==0.9.0
h2==3.2.0
hexbytes==1.2.1
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib_resources==6.3.1
kazoo==2.10.0
llvmlite==0.42.0
MouseInfo==0.1.3
msgpack==1.1.0
multidict==6.1.0
numba==0.59.0
numpy==1.26.4
oauth2client==4.1.3
openpyxl==3.1.4
packaging==23.2
parsimonious==0.10.0
pefile==2023.2.7
Pillow==10.1.0
platformdirs==4.2.0
promise==2.3
propcache==0.2.0
proto-plus==1.23.0
protobuf==3.19.6
psutil==5.9.7
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.20.0
pycryptodome-test-vectors==1.0.14
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.10
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==1.0.1
python-consul==1.1.0
python-dotenv==1.0.0
pytweening==1.2.0
pyunormalize==16.0.0
pyzmq==26.2.0
qrcode==7.4.2
ratelim==0.1.6
regex==2024.11.6
requests==2.31.0
rfc3986==1.5.0
rlp==4.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
tenacity==5.1.5
tensorflow-datasets==4.8.3
tensorflow-metadata==1.13.0
termcolor==2.4.0
thrift==0.20.0
toml==0.10.2
toolz==1.0.0
tqdm==4.66.2
typeguard==2.13.3
types-requests==2.32.0.20241016
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.1
web3==7.5.0
websockets==14.0
whichcraft==0.6.1
wrapt==1.16.0
yarl==1.17.1
zerorpc==0.6.3
zipp==3.18.1
zope.event==5.0
zope.interface==6.1
